Add an Accommodation checkbox to the form in a CampDetails component
An administrator can now set if an accommodation is being offered with
this camp. If it is not offered, an Accommodation column is not shown
on the Participants table and Accommodation checkbox is not shown on
a registration form in a CampRegistrationOptionChooserModal component.

For camp details form, a vertical space between Sayonara meeting and
Accommodation options has been changed (decreased) by moving both options
into one Form.Group tag.

For registration form, a vertical space between options has been changed
(decreased) by moving them into one Form.Group tag.