Implement advanced payment col for camps, db null protection
When a new boolean column was added to the database, it was originally filled up
with null values. However, if the boolean field filled with null comes into the
Edit form, it causes its unexceptional behavior. A protection has been
implemented, where the null boolean values received from the database are translated
to the false boolean values. It's implemented in the loadItemToUpdate() function
in setState call. The ternary operator returns true if the true value comes from
the database, but when a false or null comes - it returns false. Then the state
is updated by the setState() function.